-Interpreter: Add cast (remove warning) -AcpiGabi:  *new Device name  *comment user mode API  *Update API struct
git-tfs-id: [https://g02deas01tfoun.g02.fujitsu.local/tfs/tpc1]$/GABI_ACPI_DRIVER;C60555

+++ _DEV/Source/Drivers/FscGabi/InterpreterParse.cpp
@@ -1071,7 +1071,7 @@ eInterpreterReturn ParseModRM(psInterpreterContext pContext, PUSHORT pwCurrentIn
 				return INTERPRETER_E_EOF;
 			}
 
-			pInstruction->Displacement += (pContext->pEntryPoint[*pwCurrentIndex] << (i * 8));
+			pInstruction->Displacement += ((ULONGLONG)pContext->pEntryPoint[*pwCurrentIndex] << (i * 8));
 		}
 
 		if (sibRequired || displacementBytes > 0)
